requirements.txt
streamlit 
hugchat 
numpy 
opencv-python 
opencv-python-headless
pillow